Work Page - CSS bug fixes
Mobile & Tablet (primarily, 576px Width - 991px Width)

The Video/WEBP File at the top should be full width. Some browser/phone widths will show too much empty space on either side.  - DONE
The Work posts sometimes overlap the video at the top as well. Can you help us prevent this from happening? - Video has been hidden in all the screens, hence did not work on it

In some sizes, the copy in the "Filter" on rollover is not fitting in the frame. - I could not find this issue, please explain me or give me the screenshot of this

On Click of a Work Post opens a popup modal. This modal doesn't match the design in the mobile sizes as it's too big and too much empty space. The smaller the screen width the less it matches the design. Can your team also look into providing solves for this? - DONE